requirements.txt
rich
speedtest-cli
requests
beautifulsoup4
networkx 